fix: 상수 정의 위치 변경
