+++ src/entities/attendance/model/index.ts
@@ -1,10 +1,3 @@
 export * from "./mappers";
 
-export const ATTENDANCE_STATUS_LABELS = {
-  present: "출석",
-  late: "지각",
-  leftEarly: "조퇴",
-  leave: "공가",
-  annual: "월차",
-  absent: "결석",
-} as const;
+export * from "./constants";
